Fix all StyleCop warnings in HolidayExtractorChs
- File has been renamed to 'ChineseHolidayExtractorConfiguration' to match class name.
- Added spaces when needed
